// src/types/app.interface.ts
export interface ICharactersResponse {
    info: {
        count: number;
        pages: number;
        next: string | null;
        prev: string | null;
    }
    results: ICharacter[];
}

export interface ICharacter {
    id: number;
    name: string;
    status: string;
    species: string;
    type: string;
    gender: string;
    origin: IOrigin;
    location: ILocation;
    image: string;
    episode: string[];
    url: string;
    created: string;
}

interface IOrigin {
    name: string;
    url: string;
}

interface ILocation {
    name: string;
    url: string;
}